+++ NEWS.rst
@@ -4,6 +4,8 @@ Changes in 3.15.4 (released ??/??/2010)
 *	:mod:`ll.sisyphus` logging of exceptions and tracebacks should be more robust
 	against encoding problems.
 
+*	The :mod:`cssutils` version has been bumped to 0.9.7.
+
 
 Changes in 3.15.3 (released 11/26/2010)
 ---------------------------------------